requirements.txt
fastapi
uvicorn
confluent_kafka
pandas
python-dotenv
plotly